# perf(ext/ffi): Optimise common pointer related APIs